requirements.txt
pygame
matplotlib
pandas
pynput
numpy
pytest
scipy
websocket-client
opencv-contrib-python
sklearn
ray
redis
dynamixel_sdk
